+++ projects/ngx-markdown-editor/src/lib/config/default-editor-config.ts
@@ -1,10 +1,12 @@
-import { MarkdownEditorConfig } from "../types/config.interface";
+import { MarkdownEditorConfig } from '../types/config.interface';
 
-export const DEFAULT_EDITOR_CONFIG: MarkdownEditorConfig = {
+export const DEFAULT_EDITOR_CONFIG: Required<MarkdownEditorConfig> = {
   showPreview: true,
   showToolbar: true,
   syncScroll: true,
   value: '',
   readonly: false,
   disableSanitizer: false,
+  useDefaultStyles: true,
+  customStylesClass: '',
 };